pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Sport system backend"
authors = ["Kiselev Vladimir <wsxxx2016@yandex.ru>"]


[tool.taskipy.tasks]
manage = "python ./backend/manage.py"
runserver = "python ./backend/manage.py runserver"
migrate = "python ./backend/manage.py migrate"
makemigrations = "python ./backend/manage.py makemigrations"


[tool.black]
line-length = 79


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
line_length = 79


[tool.poetry.dependencies]
python = "3.10.10"
django = "^3.0.3"
django-environ = "^0.4.5"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.7.0"
drf-yasg = "^1.20.0"
django-constance = { extras = ["redis"], version = "^2.8.0" }
django-debug-toolbar = "^3.2"
djoser = "^2.1.0"
djangorestframework-camel-case = "^1.2.0"
djangorestframework-simplejwt = "^4.6.0"
psycopg2 = "^2.8.6"
pydantic = "^1.8.1"
django-rest-swagger = "^2.2.0"
transliterate = "^1.10.2"
django-click = "^2.2.0"
tqdm = "^4.62.2"
django-filter = "^2.4.0"
Pillow = "^8.3.2"
django-extensions = "^3.1.3"
pyparsing = "^2.4.7"
pydot = "^1.4.2"
psycopg2-binary = "^2.9.3"
django-model2puml = "^0.2.1"
django-service-objects = "^0.7.1"
django-import-export = "^3.3.1"
py-linq = "^1.4.0"


[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
autopep8 = "^1.5"
pylint-django = "^2.0.13"
taskipy = "^1.1.3"
flake8 = "^3.7.9"
black = "^20.8b1"
